requirements.txt
asgiref==3.5.2
async-generator==1.10
attrs==23.1.0
beautifulsoup4==4.12.2
certifi==2022.12.7
cffi==1.15.1
Django==4.0.5
django-bootstrap4==23.1
djangorestframework==3.14.0
exceptiongroup==1.1.1
h11==0.14.0
idna==3.4
outcome==1.2.0
psycopg2==2.9.6
pycparser==2.21
PySocks==1.7.1
pytz==2023.3
selenium==4.9.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.2
trio==0.22.0
trio-websocket==0.10.2
tzdata==2022.1
urllib3==1.26.15
wsproto==1.2.0
